DD_mall/requirements.txt
Django>=3.2.3,<3.3
django_redis
djangorestframework==3.12.2
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
asgiref==3.8.1
PyMySQL==1.1.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
uWSGI>=2.0.19.1,<2.1
djangorestframework-jwt
